CFP2-113: Extract functions to service files
