+++ src/components/person/model/person.model.ts
@@ -29,11 +29,3 @@ export const emptyPerson: PersonModel = {
     roles: [],
     createdBy: '',
 };
-
-// Helper functions
-export const getPersonDisplayNameForFormSelectFields = (person: PersonModel) => {
-    if (person.id === '') {
-        return '';
-    }
-    return `${person.firstName} ${person.lastName} (${person.address.city})`;
-};
