+++ src/components/property/service/property.service.ts
@@ -1,0 +1,5 @@
+import { PropertyModel } from '../model/property.model';
+
+export const getDisplayNameOfProperty = (p: PropertyModel) => {
+    return `${p.name} (${p.egid})`;
+};
