+++ src/components/rental-unit/overview/rental-units-table.tsx
@@ -15,11 +15,13 @@ import { Link } from 'react-router-dom';
 import routes from '../../../routes/route-constants';
 import EditOutlinedIcon from '@material-ui/icons/EditOutlined';
 import DeleteOutlineIcon from '@material-ui/icons/DeleteOutline';
-import { getDisplayNameOfRentalUnit, RentalUnitModel } from '../model/rental-unit.model';
+import { RentalUnitModel } from '../model/rental-unit.model';
 import { useDeletePrompt } from '../../../hooks/ui.hooks';
-import { getTenantsOfTenancy, TenancyModel } from '../../tenancy/model/tenancy.model';
+import { TenancyModel } from '../../tenancy/model/tenancy.model';
 import { parseISO } from 'date-fns';
 import { PersonModel } from '../../person/model/person.model';
+import * as rentalUnitService from '../service/rental-unit.service';
+import * as tenancyService from '../../tenancy/service/tenancy.service';
 
 type RentalUnitsTableProps = {
     tenancies: TenancyModel[];
@@ -46,7 +48,7 @@ export const RentalUnitsTable = ({
         useDeletePrompt();
 
     const getTenantsDesc = (ruId: string) => {
-        const ts = getTenantsOfTenancy(
+        const ts = tenancyService.getTenantsOfTenancy(
             // TODO reuse filters and finds
             tenancies
                 ?.filter((t) => !t.endOfContract || parseISO(t.endOfContract) > new Date())
@@ -99,7 +101,9 @@ export const RentalUnitsTable = ({
                                         <DeleteOutlineIcon color={'error'} />
                                     </IconButton>
                                 </TableCell>
-                                <TableCell align="right">{getDisplayNameOfRentalUnit(ru)}</TableCell>
+                                <TableCell align="right">
+                                    {rentalUnitService.getDisplayNameOfRentalUnit(ru)}
+                                </TableCell>
                                 {/* // TODO: Make font bold and red (error) if it is a Vacancy */}
                                 <TableCell align="right">{getTenantsDesc(ru.id)}</TableCell>
                             </TableRow>
