+++ src/components/tenancy/model/tenancy.model.ts
@@ -1,5 +1,4 @@
 import { months } from '../../../constants';
-import { PersonModel } from '../../person/model/person.model';
 import { format } from 'date-fns';
 
 export type TenancyModel = {
@@ -31,31 +30,4 @@ export const emptyTenancy: TenancyModel = {
     utilities: 0,
     rentDeposit: 0,
     rentAccount: '',
-};
-
-// Helper functions
-export const getTenantsOfTenancy = (tenancy: TenancyModel | undefined, allTenants: PersonModel[]) => {
-    if (!tenancy) {
-        return [];
-    }
-
-    return allTenants.filter((t) => t.id === tenancy.tenant1Id || t.id === tenancy.tenant2Id);
-};
-
-export const getDisplayNameOfTenants = (tenancy: TenancyModel, allTenants: PersonModel[]) => {
-    if (tenancy === undefined) return '-';
-    return getTenantsOfTenancy(tenancy, allTenants)
-        .map((t) => t.firstName + ' ' + t.lastName)
-        .join(', ');
-};
-
-export const getDisplayNameOfTenancy = (tenancy: TenancyModel, allTenants: PersonModel[]) => {
-    const tenants = getTenantsOfTenancy(tenancy, allTenants);
-    const desc = 'Tenancy';
-
-    if (!tenants?.length) {
-        return desc + ' (Vacancy)';
-    }
-
-    return `${desc} (${tenants.map((t) => t.firstName + ' ' + t.lastName).join(' & ')})`;
 };
