+++ src/components/tenancy/tenancy-detail.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { emptyTenancy, getDisplayNameOfTenancy } from './model/tenancy.model';
+import { emptyTenancy } from './model/tenancy.model';
 import { TenancyForm } from './tenancy-form';
 import { useParams } from 'react-router';
 import { useAppSelector } from '../../hooks/store.hooks';
@@ -12,6 +12,7 @@ import {
 import { emptyProperty } from '../property/model/property.model';
 import { emptyRentalUnit } from '../rental-unit/model/rental-unit.model';
 import { IntroHeader } from '../ui/intro-header/intro-header';
+import * as tenancyService from './service/tenancy.service';
 
 export const TenancyDetail = ({ isNew }: { isNew: boolean }) => {
     const { id } = useParams<{ id: string }>();
@@ -37,7 +38,7 @@ export const TenancyDetail = ({ isNew }: { isNew: boolean }) => {
                     isNew
                         ? 'Create a new tenancy'
                         : tenancyToEdit
-                        ? getDisplayNameOfTenancy(tenancyToEdit, tenants)
+                        ? tenancyService.getDisplayNameOfTenancy(tenancyToEdit, tenants)
                         : 'Tenancy has not been found!'
                 }
             />
